NONRTRIC - A scheme for versioning of types.
If a consumer creates a job and the requested type does not exist, a compatible type may be automatically chosen.
The type with the lowest version number is chosen.

All producers that has been registerred to support a compatible type are notified.
The producers will one of its registerred types (the one with the lowest number).

Change-Id: I94580f3fd1e07c5b0483232e06d1950c2c573fea
Signed-off-by: PatrikBuhr <patrik.buhr@est.tech>
Issue-ID: NONRTRIC-810