Fix Unit test logging failure
Spring Boot Version 3.2.0(Last used before upgrade) contains logback library version 1.4.11.
logback version 1.4.11 is vulnerable to CVE-2023-6378. This CVE is related to creating missing parent directories.
It is fixed as part of 1.4.12+ (https://github.com/advisories/GHSA-vmq6-5m68-f53m). The fix is to avoid creating missing parent directories.

We are getting couple of reports related to this issue in the ApplicationTest unit test(After Spring Boot upgrade to 3.5.5 (Which includes logback 1.5.18)).

So the fix is to introduce a temporary application log file for the unit test.

This doesn't happen when running this container because the /var/log/information-coordinator-service is created as part of the Docker build (In Dockerfile).

Issue-ID: INT-193
Change-Id: I63088541045b3faf0038223c4e81f7ab35d7a878
Signed-off-by: aravind.est <aravindhan.a@est.tech>